Improve checkstyle
