+++ src/test/java/seedu/address/model/calorie/DailyCalorieTest.java
@@ -1,188 +1,72 @@
 package seedu.address.model.calorie;
 
-import org.junit.jupiter.api.Test;
-import seedu.address.model.person.CalorieLog;
-import seedu.address.model.person.DailyCalorie;
-import seedu.address.model.person.exceptions.DailyCalorieNotFoundException;
-import seedu.address.model.person.exceptions.DuplicateDailyCalorieException;
-
-import java.time.LocalDate;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.testutil.Assert.assertThrows;
+
+import java.time.LocalDate;
+
+import org.junit.jupiter.api.Test;
+
+import seedu.address.model.person.DailyCalorie;
 
 public class DailyCalorieTest {
 
-    LocalDate testDate1 = LocalDate.of(2020, 10, 28);
-    LocalDate testDate2 = LocalDate.of(2020, 10, 26);
-    DailyCalorie d1 = new DailyCalorie(testDate1);
-    DailyCalorie d2 = new DailyCalorie(testDate1);
-    DailyCalorie d3 = new DailyCalorie(testDate2);
-    CalorieLog l1 = new CalorieLog();
-    CalorieLog l2 = new CalorieLog();
+    private final LocalDate testDate1 = LocalDate.of(2020, 10, 28);
+    private final DailyCalorie d1 = new DailyCalorie(testDate1);
+    private final DailyCalorie d2 = new DailyCalorie(testDate1);
+
 
     @Test
-    public void equals_DifferentCalorieLog_returnsTrue() {
-        assertTrue(l1.equals(l2));
+    public void isSameDailyCalorie() {
+        // same object -> returns true
+        assertTrue(d1.isSameDailyCalorie(d1));
+        assertTrue(d2.isSameDailyCalorie(d2));
+
+        // null -> returns false
+        assertFalse(d1.isSameDailyCalorie(null));
+        assertFalse(d2.isSameDailyCalorie(null));
+
+        // same LocalDate -> returns true
+        assertTrue(d1.isSameDailyCalorie(d2));
+
+        // same LocalDate, different calories -> returns true
+        d1.addCalories(10);
+        assertTrue(d1.isSameDailyCalorie(d2));
+
+        // same LocalDate, same calories -> returns true
+        d2.addCalories(10);
+        assertTrue(d1.isSameDailyCalorie(d2));
+
     }
 
     @Test
-    public void equals_DifferentCalorieLogWithDifferentDailyCalorie_returnsFalse() {
-        l1.add(d1);
-        assertFalse(l1.equals(l2));
-    }
+    public void equals() {
+        LocalDate testDate2 = LocalDate.of(2020, 10, 26);
+        DailyCalorie d3 = new DailyCalorie(testDate2);
 
-    @Test
-    public void equals_DifferentCalorieLogWithSameDailyCalorie_returnsTrue() {
-        l1.add(d1);
-        l2.add(d1);
-        assertTrue(l1.equals(l2));
-    }
+        // same values with no calories -> returns true
+        assertTrue(d1.equals(d2));
 
-    @Test
-    public void contains_nullCalorieLog_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> l1.contains(null));
-    }
+        // same object -> returns true
+        assertTrue(d1.equals(d1));
+        assertTrue(d2.equals(d2));
 
-    @Test
-    public void contains_DailyCalorieNotInList_returnsFalse() {
-        assertFalse(l1.contains(d1));
-    }
+        // null -> returns false
+        assertFalse(d1.equals(null));
+        assertFalse(d2.equals(null));
 
-    @Test
-    public void contains_DailyCalorieInList_returnsTrue() {
-        l1.add(d1);
-        assertTrue(l1.contains(d1));
-    }
+        // different type -> returns false
+        assertFalse(d1.equals(5));
 
-    @Test
-    public void contains_DailyCalorieWithSameIdentityFieldsInList_returnsTrue() {
-        l1.add(d1);
-        assertTrue(l1.contains(d2));
-    }
+        // same LocalDate, different DailyCalorie -> returns false
+        d2.addCalories(1500);
+        assertFalse(d1.equals(d2));
 
-    @Test
-    public void add_nullDailyCalorie_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> l1.add(null));
-    }
+        // different LocalDate, same DailyCalorie -> returns false
+        assertFalse(d1.equals(d3));
 
-    @Test
-    public void add_duplicateDailyCalorie_throwsDuplicateDailyCalorieException() {
-        l1.add(d1);
-        assertThrows(DuplicateDailyCalorieException.class, () -> l1.add(d1));
-    }
-
-    @Test
-    public void setDailyCalorie_nullTargetDailyCalorie_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> l1.setDailyCalorie(null, d1));
-    }
-
-    @Test
-    public void setDailyCalorie_nullEditedDailyCalorie_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> l1.setDailyCalorie(d1, null));
-    }
-
-    @Test
-    public void setDailyCalorie_targetDailyCalorieNotInList_throwsDailyCalorieNotFoundException() {
-        assertThrows(DailyCalorieNotFoundException.class, () -> l1.setDailyCalorie(d1, d1));
-    }
-
-    @Test
-    public void setDailyCalorie_editedDailyCalorieIsSameDailyCalorie_success() {
-        l1.add(d1);
-        l1.setDailyCalorie(d1, d1);
-        CalorieLog expectedCalorieLog = new CalorieLog();
-        expectedCalorieLog.add(d1);
-        assertEquals(l1, expectedCalorieLog);
-    }
-
-    @Test
-    public void setDailyCalorie_editedDailyCalorieHasSameIdentity_success() {
-        l1.add(d1);
-        l1.setDailyCalorie(d1, d2);
-        CalorieLog expectedCalorieLog = new CalorieLog();
-        expectedCalorieLog.add(d2);
-        assertEquals(l1, expectedCalorieLog);
-    }
-
-    @Test
-    public void setDailyCalorie_editedDailyCalorieHasDifferentIdentity_success() {
-        l1.add(d1);
-        l1.setDailyCalorie(d1, d2);
-        CalorieLog expectedCalorieLog = new CalorieLog();
-        expectedCalorieLog.add(d2);
-        assertEquals(l1, expectedCalorieLog);
-    }
-
-    @Test
-    public void setDailyCalorie_editedDailyCalorieHasNonUniqueIdentity_throwsDuplicateDailyCalorieException() {
-        l1.add(d1);
-        l1.add(d3);
-        assertThrows(DuplicateDailyCalorieException.class, () -> l1.setDailyCalorie(d1, d3));
-    }
-
-    @Test
-    public void remove_nullDailyCalorie_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> l1.remove(null));
-    }
-
-    @Test
-    public void remove_dailyCalorieDoesNotExist_throwsDailyCalorieNotFoundException() {
-        assertThrows(DailyCalorieNotFoundException.class, () -> l1.remove(d1));
-    }
-
-    @Test
-    public void remove_existingDailyCalorie_removesDailyCalorie() {
-        l1.add(d1);
-        l1.remove(d1);
-        CalorieLog expectedCalorieLog = new CalorieLog();
-        assertEquals(expectedCalorieLog, l1);
-    }
-
-    @Test
-    public void setDailyCalories_nullUniqueDailyCalorieList_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> l1.setDailyCalories((CalorieLog) null));
-    }
-
-    @Test
-    public void setDailyCalories_uniqueDailyCalorieList_replacesOwnListWithProvidedCalorieLog() {
-        l1.add(d1);
-        CalorieLog expectedCalorieLog = new CalorieLog();
-        expectedCalorieLog.add(d2);
-        l1.setDailyCalories(expectedCalorieLog);
-        assertEquals(l1, expectedCalorieLog);
-    }
-
-    @Test
-    public void setDailyCalories_nullList_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> l1.setDailyCalories((List<DailyCalorie>) null));
-    }
-
-    @Test
-    public void setDailyCalories_list_replacesOwnListWithProvidedList() {
-        l1.add(d1);
-        List<DailyCalorie> dailyCalorieList = Collections.singletonList(d1);
-        l1.setDailyCalories(dailyCalorieList);
-        CalorieLog expectedCalorieLog = new CalorieLog();
-        expectedCalorieLog.add(d1);
-        assertEquals(expectedCalorieLog, l1);
-    }
-
-    @Test
-    public void setDailyCalories_listWithDuplicateDailyCalories_throwsDuplicateDailyCalorieException() {
-        List<DailyCalorie> listWithDuplicateDailyCalories = Arrays.asList(d1, d1);
-        assertThrows(DuplicateDailyCalorieException.class, () -> l1.setDailyCalories(listWithDuplicateDailyCalories));
-    }
-
-    @Test
-    public void asUnmodifiableObservableList_modifyList_throwsUnsupportedOperationException() {
-        assertThrows(UnsupportedOperationException.class, () -> l1
-                .asUnmodifiableObservableList().remove(0));
+        // different LocalDate, different DailyCalorie -> returns false
+        assertFalse(d2.equals(d3));
     }
 
 }
